+++ controllers/home-routes.js
@@ -7,6 +7,11 @@ router.get('/', (req, res) => {
 });
 
 router.get('/login', (req, res) => {
+    // if (req.session.loggedIn) {
+    //     res.redirect('/');
+    //     return;
+    // }
+    
     res.render('login');
 });
 module.exports = router;
