+++ controllers/index.js
@@ -1,8 +1,8 @@
 const router = require('express').Router();
-//const apiRoutes = require('./api/');
+const apiRoutes = require('./api/');
 const homeRoutes = require('./home-routes.js');
 //const dashboardRoutes = require('./dashboard-routes.js');
 router.use('/', homeRoutes);
 //router.use('/dashboard', dashboardRoutes);
-//router.use('/api', apiRoutes);
+router.use('/api', apiRoutes);
 module.exports = router;
