+++ public/javascript/button-login.js
@@ -1,0 +1,14 @@
+function signupRedirect(event) {
+  event.preventDefault();
+
+  console.log("works")
+  window.location.href = '/login';
+}
+
+function loginRedirect(event) {
+  event.preventDefault();
+  window.location.href = '/login';
+}
+
+document.getElementById('signupButton').addEventListener('click', signupRedirect);
+document.getElementById('loginButton').addEventListener('click', loginRedirect);
