+++ public/javascript/login.js
@@ -1,19 +1,4 @@
 
-function signupRedirect(event) {
-  event.preventDefault();
-
-  console.log("works")
-  window.location.href = '/login';
-
-
-}
-
-function loginRedirect(event) {
-  event.preventDefault();
-  window.location.href = '/login';
-
-}
-
 
 
 async function loginHandler(event) {
@@ -21,7 +6,7 @@ async function loginHandler(event) {
   const email = document.querySelector('#emailLogin').value.trim();
   const password = document.querySelector('#passwordLogin').value.trim();
   if (email && password) {
-    const response = await fetch('/api/users/login', {
+    const response = await fetch('/api/user/login', {
       method: 'post',
       body: JSON.stringify({
         email,
@@ -43,7 +28,7 @@ async function signupHandler(event) {
   const email = document.querySelector('#emailSignup').value.trim();
   const password = document.querySelector('#passwordSignup').value.trim();
   if (username && email && password) {
-    const response = await fetch('/api/users', {
+    const response = await fetch('/api/user', {
       method: 'post',
       body: JSON.stringify({
         username,
@@ -62,7 +47,6 @@ async function signupHandler(event) {
 }
 
 
-document.getElementById('signupButton').addEventListener('click', signupRedirect);
-document.getElementById('loginButton').addEventListener('click', loginRedirect);
+
 document.querySelector('.signupForm').addEventListener('submit', signupHandler);
 document.querySelector('.loginForm').addEventListener('submit', loginHandler);
