+++ views/layouts/main.handlebars
@@ -137,4 +137,4 @@
 </html>
 
 
-<script src="../javascript/login.js"></script>
+<script src="../javascript/button-login.js"></script>
